combine `BusyStateHelper`, `ErrorStateHelper`, `FutureRunnerHelper`
