+++ packages/stacked/example/lib/ui/smart_widgets/widget_one/widget_one_viewmodel.dart
@@ -2,7 +2,7 @@ import 'package:example/app/app.locator.dart';
 import 'package:example/services/information_service.dart';
 import 'package:stacked/stacked.dart';
 
-class WidgetOneViewModel extends ReactiveViewModel with FutureRunnerHelper {
+class WidgetOneViewModel extends ReactiveViewModel {
   final InformationService _informationService =
       exampleLocator<InformationService>();
   int get postCount => _informationService.postCount;
