+++ packages/stacked/lib/stacked.dart
@@ -18,7 +18,6 @@ export 'src/state_management/form_view_model.dart';
 export 'src/state_management/index_tracking_viewmodel.dart';
 
 /// viewmodel helpers
-export 'src/state_management/helpers/future_runner_helper.dart';
 export 'src/state_management/helpers/message_state_helper.dart';
 export 'src/state_management/helpers/data_state_helper.dart';
 export 'src/state_management/helpers/index_tracking_state_helper.dart';
